Added spinner when getting data + moved symptom service
