+++ client/src/app/pages/user-data-edit-page/user-data-edit-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { DropdownSelectComponent } from 'src/app/components/dropdown-select/dropdown-select.component';
-import { SymptomsAPIService } from '../../services/symptomsAPI/symptoms-api.service';
+import { SymptomsAPIService } from '../../services/symptoms-api.service';
 
 @Component({
   selector: 'app-user-data-edit-page',
