+++ client/src/app/services/symptoms-api.service.ts
